#31 sop exercisecard and button functionalities
